# internal/web/tasks/api.gen.go
// Package tasks provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package tasks

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// Task defines model for Task.
type Task struct {
	Id     *string `json:"id,omitempty"`
	IsDone *bool   `json:"is_done,omitempty"`
	Text   *string `json:"text,omitempty"`
	UserId *string `json:"user_id,omitempty"`
}

// PatchTasksTaskIdJSONRequestBody defines body for PatchTasksTaskId for application/json ContentType.
type PatchTasksTaskIdJSONRequestBody = Task

// PostTasksUserIdJSONRequestBody defines body for PostTasksUserId for application/json ContentType.
type PostTasksUserIdJSONRequestBody = Task

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get all tasks
	// (GET /tasks)
	GetTasks(ctx echo.Context) error
	// Delete a task
	// (DELETE /tasks/{task_id})
	DeleteTasksTaskId(ctx echo.Context, taskId string) error
	// Update a task
	// (PATCH /tasks/{task_id})
	PatchTasksTaskId(ctx echo.Context, taskId string) error
	// Create a new task
	// (POST /tasks/{user_id})
	PostTasksUserId(ctx echo.Context, userId string) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetTasks converts echo context to params.
func (w *ServerInterfaceWrapper) GetTasks(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetTasks(ctx)
	return err
}

// DeleteTasksTaskId converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteTasksTaskId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "task_id" -------------
	var taskId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "task_id", runtime.ParamLocationPath, ctx.Param("task_id"), &taskId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter task_id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteTasksTaskId(ctx, taskId)
	return err
}

// PatchTasksTaskId converts echo context to params.
func (w *ServerInterfaceWrapper) PatchTasksTaskId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "task_id" -------------
	var taskId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "task_id", runtime.ParamLocationPath, ctx.Param("task_id"), &taskId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter task_id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchTasksTaskId(ctx, taskId)
	return err
}

// PostTasksUserId converts echo context to params.
func (w *ServerInterfaceWrapper) PostTasksUserId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "user_id" -------------
	var userId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "user_id", runtime.ParamLocationPath, ctx.Param("user_id"), &userId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter user_id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostTasksUserId(ctx, userId)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/tasks", wrapper.GetTasks)
	router.DELETE(baseURL+"/tasks/:task_id", wrapper.DeleteTasksTaskId)
	router.PATCH(baseURL+"/tasks/:task_id", wrapper.PatchTasksTaskId)
	router.POST(baseURL+"/tasks/:user_id", wrapper.PostTasksUserId)

}

type GetTasksRequestObject struct {
}

type GetTasksResponseObject interface {
	VisitGetTasksResponse(w http.ResponseWriter) error
}

type GetTasks200JSONResponse []Task

func (response GetTasks200JSONResponse) VisitGetTasksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type DeleteTasksTaskIdRequestObject struct {
	TaskId string `json:"task_id"`
}

type DeleteTasksTaskIdResponseObject interface {
	VisitDeleteTasksTaskIdResponse(w http.ResponseWriter) error
}

type DeleteTasksTaskId204Response struct {
}

func (response DeleteTasksTaskId204Response) VisitDeleteTasksTaskIdResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type PatchTasksTaskIdRequestObject struct {
	TaskId string `json:"task_id"`
	Body   *PatchTasksTaskIdJSONRequestBody
}

type PatchTasksTaskIdResponseObject interface {
	VisitPatchTasksTaskIdResponse(w http.ResponseWriter) error
}

type PatchTasksTaskId200JSONResponse Task

func (response PatchTasksTaskId200JSONResponse) VisitPatchTasksTaskIdResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostTasksUserIdRequestObject struct {
	UserId string `json:"user_id"`
	Body   *PostTasksUserIdJSONRequestBody
}

type PostTasksUserIdResponseObject interface {
	VisitPostTasksUserIdResponse(w http.ResponseWriter) error
}

type PostTasksUserId201JSONResponse Task

func (response PostTasksUserId201JSONResponse) VisitPostTasksUserIdResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Get all tasks
	// (GET /tasks)
	GetTasks(ctx context.Context, request GetTasksRequestObject) (GetTasksResponseObject, error)
	// Delete a task
	// (DELETE /tasks/{task_id})
	DeleteTasksTaskId(ctx context.Context, request DeleteTasksTaskIdRequestObject) (DeleteTasksTaskIdResponseObject, error)
	// Update a task
	// (PATCH /tasks/{task_id})
	PatchTasksTaskId(ctx context.Context, request PatchTasksTaskIdRequestObject) (PatchTasksTaskIdResponseObject, error)
	// Create a new task
	// (POST /tasks/{user_id})
	PostTasksUserId(ctx context.Context, request PostTasksUserIdRequestObject) (PostTasksUserIdResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetTasks operation middleware
func (sh *strictHandler) GetTasks(ctx echo.Context) error {
	var request GetTasksRequestObject

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetTasks(ctx.Request().Context(), request.(GetTasksRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTasks")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetTasksResponseObject); ok {
		return validResponse.VisitGetTasksResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// DeleteTasksTaskId operation middleware
func (sh *strictHandler) DeleteTasksTaskId(ctx echo.Context, taskId string) error {
	var request DeleteTasksTaskIdRequestObject

	request.TaskId = taskId

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteTasksTaskId(ctx.Request().Context(), request.(DeleteTasksTaskIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteTasksTaskId")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(DeleteTasksTaskIdResponseObject); ok {
		return validResponse.VisitDeleteTasksTaskIdResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PatchTasksTaskId operation middleware
func (sh *strictHandler) PatchTasksTaskId(ctx echo.Context, taskId string) error {
	var request PatchTasksTaskIdRequestObject

	request.TaskId = taskId

	var body PatchTasksTaskIdJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PatchTasksTaskId(ctx.Request().Context(), request.(PatchTasksTaskIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PatchTasksTaskId")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PatchTasksTaskIdResponseObject); ok {
		return validResponse.VisitPatchTasksTaskIdResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PostTasksUserId operation middleware
func (sh *strictHandler) PostTasksUserId(ctx echo.Context, userId string) error {
	var request PostTasksUserIdRequestObject

	request.UserId = userId

	var body PostTasksUserIdJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostTasksUserId(ctx.Request().Context(), request.(PostTasksUserIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostTasksUserId")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PostTasksUserIdResponseObject); ok {
		return validResponse.VisitPostTasksUserIdResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}